Initial import, although this probably isn't yet stable enough to be used on a production site
MiniInvite is a stripped-down version of the InviteContacts extension, and
it doesn't include the actual contact importing functionality, mainly
because contact importing is a pain. Some e-mail providers don't provide
an API, some sites explicitly forbid screenscraping in their ToU and even
ignoring those facts, the contact importing libraries I tested just didn't
work.

The hooked functions in InviteFriendOnEdit.php are kinda weird and scary.
Some of the code has been salvaged from old skin files in the ArmchairGM
codebase and hard-coding the background color for span#invite_to_edit &
span#invite_to_read is not ideal and will be fixed in the future.
Besides that, there's a hard-coded category name on line 45, but I'm not
sure what the "proper" fix here is, given recent BlogPage changes, and I'm
afraid it'll involve a lot of scary regex voodoo.

But hey, at least the code mostly works and is open, so yay!

Change-Id: I6b2f87ea25cd3f0902dcd801c3b1e202ffef7554